+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ### Changed
 
+- `require "kettle/test"` no longer loads `version_gem` by default; require
+  `kettle/test/version_gem` for the optional `VersionGem::Basic` extension.
+
 ### Deprecated
 
 ### Removed